+++ R/head.piar_index.R
@@ -5,7 +5,7 @@
 #' @param x A price index, as made by, e.g., [elemental_index()].
 #' @param n See [head()]/[tail()]. The default takes the
 #' first/last 6 levels of `x`.
-#' @param ... Not used.
+#' @param ... Not currently used.
 #'
 #' @returns
 #' A price index that inherits from [`chainable_piar_index`] if `x` is a